+++ packages/core/src/index.d.ts
@@ -33,7 +33,7 @@ export interface PageLayout {
         scrollMode: ScrollMode;
         viewMode: ViewMode;
     }) => React.CSSProperties;
-    tranformSize?: ({ numPages, pageIndex, size }: { numPages: number; pageIndex: number; size: Rect }) => Rect;
+    transformSize?: ({ numPages, pageIndex, size }: { numPages: number; pageIndex: number; size: Rect }) => Rect;
 }
 export interface PageSize {
     pageHeight: number;